move graph classes into hashgraph.cc,hh
